+++ example/example.dart
@@ -8,7 +8,7 @@ void main(List<String> args) {
   createProject(
     Project(
       name: 'folder name',
-      version: 17,
+      version: 18,
       target: './', // path for where to generate the project
       generate: ExamplePack(), // The starting point of generation
     ),